Create a new ComboBox extension.
